location_api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic-settings==2.5.2
SQLAlchemy==2.0.36
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.2
python-dotenv==1.0.1
fastapi-cache2==0.2.2
redis==5.0.8
orjson==3.10.7
httpx==0.27.2
pytest==8.3.2
pytest-asyncio==0.23.8
starlette==0.38.6
strawberry-graphql==0.235.1
geoalchemy2==0.15.2
shapely==2.0.6
